bölme işlemi hata giderildi.

+++ Program.cs
@@ -7,9 +7,9 @@
             Console.WriteLine("Hesap Makinesi\n");
 
             Console.Write("Birinci Sayıyı Giriniz: ");
-            int sayi1 = int.Parse(Console.ReadLine());
+            double sayi1 = int.Parse(Console.ReadLine());
             Console.Write("İkinci Sayı Giriniz: ");
-            int sayi2 = int.Parse(Console.ReadLine());
+            double sayi2 = int.Parse(Console.ReadLine());
             Console.Write("İşlem Seçiniz(+, -, *, /, %): ");
             string islem = Console.ReadLine();
